backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.3
psycopg2-binary==2.9.9
djoser==2.1.0
Pillow
python-dotenv==1.0.1
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0